+++ src/main/java/com/liberty52/product/service/controller/OptionDetailCreateController.java
@@ -14,13 +14,5 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class OptionDetailCreateController {
 
-    private final OptionDetailCreateService optionDetailCreateService;
 
-    @Operation(summary = "옵션 상세 생성", description = "관리자가 새로운 옵션 상세를 생성하는 엔드포인트입니다.")
-    @PostMapping("/admin/optionDetail/{optionId}")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void createOptionDetailByAdmin(@RequestHeader("LB-Role") String role,
-                                   @Validated @RequestBody CreateOptionDetailRequestDto dto, @PathVariable String optionId) {
-        optionDetailCreateService.createOptionDetailByAdmin(role, dto, optionId);
-    }
 }
